+++ src/main/java/gregtech/api/interfaces/tileentity/IEnergyConnected.java
@@ -60,16 +60,16 @@ public interface IEnergyConnected extends IColoredTileEntity, IHasWorldObjectAnd
          */
         public static final long emitEnergyToNetwork(long aVoltage, long aAmperage, IEnergyConnected aEmitter) {
             long rUsedAmperes = 0;
-            for (byte i = 0, j = 0; i < 6 && aAmperage > rUsedAmperes; i++)
-                if (aEmitter.outputsEnergyTo(i)) {
+            for (int i = 0, j = 0; i < 6 && aAmperage > rUsedAmperes; i++)
+                if (aEmitter.outputsEnergyTo((byte)i)) {
                     j = GT_Utility.getOppositeSide(i);
-                    TileEntity tTileEntity = aEmitter.getTileEntityAtSide(i);
+                    TileEntity tTileEntity = aEmitter.getTileEntityAtSide((byte)i);
                     if (tTileEntity instanceof IEnergyConnected) {
                         if (aEmitter.getColorization() >= 0) {
                             byte tColor = ((IEnergyConnected) tTileEntity).getColorization();
                             if (tColor >= 0 && tColor != aEmitter.getColorization()) continue;
                         }
-                        rUsedAmperes += ((IEnergyConnected) tTileEntity).injectEnergyUnits(j, aVoltage, aAmperage - rUsedAmperes);
+                        rUsedAmperes += ((IEnergyConnected) tTileEntity).injectEnergyUnits((byte)j, aVoltage, aAmperage - rUsedAmperes);
                     } else if (tTileEntity instanceof IEnergySink) {
                         if (((IEnergySink) tTileEntity).acceptsEnergyFrom((TileEntity) aEmitter, ForgeDirection.getOrientation(j))) {
                             while (aAmperage > rUsedAmperes && ((IEnergySink) tTileEntity).getDemandedEnergy() > 0 && ((IEnergySink) tTileEntity).injectEnergy(ForgeDirection.getOrientation(j), aVoltage, aVoltage) < aVoltage)
